agent/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
requests
openai